requirements.txt
Flask
OpenAI
google-cloud-vision
opencv-python
flask_socketio
Pillow
python-socketio
eventlet
emit